+++ packages/admin-ui/src/layouts/index.ts
@@ -1,0 +1,3 @@
+export { SectionLayout } from "./SectionLayout";
+export type { NavItem, SectionLayoutProps } from "./SectionLayout";
+export { DecorativeBackground } from "./DecorativeBackground";
