+++ packages/admin-ui/src/pages-new/layouts/index.ts
@@ -1,2 +1,1 @@
 export { NewPageLayout } from "./NewPageLayout";
-export { DecorativeBackground } from "./DecorativeBackground";
